Fix 'Connection Is Not Secure' Firefox warning
Firefox warns with a warning sign attached to the lock icon in the URL
bar when some resources of a HTTPS page are loaded over HTTP.

In this case the offending resource is a SVG file - which is also
available over HTTPS, such this fix.

SVGs apparently may contain all kind of stuff like Javascript, thus
it's probably sensible for Firefox warning about it.